# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Audio Player API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tracks:
    get:
      operationId: listTracks
      tags: [tracks]
      summary: List all tracks
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: artist
          in: query
          schema:
            type: string
        - name: album
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortDir
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Create a track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /tracks/bulk:
    post:
      operationId: bulkUpsertTracks
      tags: [tracks]
      summary: Bulk upsert tracks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpsertTracksBody"
      responses:
        "200":
          description: Upserted tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpsertTracksResponse"
  /tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get a track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update a track (tag edit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackBody"
      responses:
        "200":
          description: Updated track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete a track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /subsonic-servers:
    get:
      operationId: listSubsonicServers
      tags: [subsonic]
      summary: List Subsonic servers
      responses:
        "200":
          description: Servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubsonicServer"
    post:
      operationId: createSubsonicServer
      tags: [subsonic]
      summary: Add a Subsonic server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubsonicServerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsonicServer"
  /subsonic-servers/{id}:
    patch:
      operationId: updateSubsonicServer
      tags: [subsonic]
      summary: Update a Subsonic server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubsonicServerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsonicServer"
    delete:
      operationId: deleteSubsonicServer
      tags: [subsonic]
      summary: Delete a Subsonic server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /subsonic-servers/{id}/test:
    get:
      operationId: testSubsonicServer
      tags: [subsonic]
      summary: Test Subsonic server connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsonicTestResult"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: ["integer", "null"]
        genre:
          type: ["string", "null"]
        duration:
          type: number
        trackNumber:
          type: ["integer", "null"]
        fileName:
          type: string
        folderPath:
          type: string
        albumArtDataUrl:
          type: ["string", "null"]
        rating:
          type: integer
        source:
          type: string
        subsonicId:
          type: ["string", "null"]
        subsonicServerId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - artist
        - album
        - duration
        - fileName
        - folderPath
        - rating
        - source
        - createdAt
        - updatedAt
    CreateTrackBody:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: ["integer", "null"]
        genre:
          type: ["string", "null"]
        duration:
          type: number
        trackNumber:
          type: ["integer", "null"]
        fileName:
          type: string
        folderPath:
          type: string
        albumArtDataUrl:
          type: ["string", "null"]
        rating:
          type: integer
        source:
          type: string
        subsonicId:
          type: ["string", "null"]
        subsonicServerId:
          type: ["integer", "null"]
      required:
        - title
        - fileName
    UpdateTrackBody:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: ["integer", "null"]
        genre:
          type: ["string", "null"]
        trackNumber:
          type: ["integer", "null"]
        rating:
          type: integer
        albumArtDataUrl:
          type: ["string", "null"]
    BulkUpsertTracksBody:
      type: object
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/CreateTrackBody"
      required:
        - tracks
    BulkUpsertTracksResponse:
      type: object
      properties:
        upserted:
          type: integer
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
      required:
        - upserted
        - tracks
...
    SubsonicServer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        username:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - url
        - username
        - isActive
        - createdAt
        - updatedAt
    CreateSubsonicServerBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        username:
          type: string
        password:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - url
        - username
        - password
    UpdateSubsonicServerBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        username:
          type: string
        password:
          type: string
        isActive:
          type: boolean
    SubsonicTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        version:
          type: ["string", "null"]
      required:
        - success
        - message