# lib/api-zod/src/generated/types/bulkUpsertTracksBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTrackBody } from "./createTrackBody";

export interface BulkUpsertTracksBody {
  tracks: CreateTrackBody[];
}