# lib/api-zod/src/generated/types/createSubsonicServerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSubsonicServerBody {
  name: string;
  url: string;
  username: string;
  password: string;
  isActive?: boolean;
}