# lib/api-zod/src/generated/types/subsonicServer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */

export interface SubsonicServer {
  id: number;
  name: string;
  url: string;
  username: string;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}