# lib/api-zod/src/generated/types/updateTrackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTrackBody {
  title?: string;
  artist?: string;
  album?: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  trackNumber?: number | null;
  rating?: number;
  /** @nullable */
  albumArtDataUrl?: string | null;
}